Create README.md
To run the project, open the RealEstateFinder project folder in Netbeans or Eclipse. We developed the application in Netbeans. Before most things can be 
performed in the application, all Properties must have an owner. Sign up as a Seller by selecting the properties the Seller owns and check that the user is 
a Seller. Whenever an Account is created, Customer or Seller, or other actions are performed on the application(Adding a Property to Favorites),
the user must click to "Log out" which is located on the navigation bar in order to save the changes to the file.